+++ internal/driver/db_structure.go
@@ -1,7 +1,7 @@
 package driver
 
 var (
-	schemaVersion = "42"
+	schemaVersion = "43"
 
 	UUID      = ColumnType{SQLite: "TEXT", MySQL: "VARCHAR(36)"}
 	DateTime  = ColumnType{SQLite: "TEXT", MySQL: "VARCHAR(32)"}
@@ -127,6 +127,7 @@ var currencyTable = NewTable("currency", "uuid", UUID,
 		{Name: "time_frame_since", Type: DateTime, Properties: "NOT NULL DEFAULT ''"},
 		{Name: "comparison", Type: SmallText, Properties: "NOT NULL"},
 		{Name: "filters", Type: FullText},
+		{Name: "sub_metrics", Type: FullText, Properties: "NOT NULL DEFAULT ''"},
 	})
 
 var customFieldsTable = NewTable("custom_fields", "uuid", UUID,