+++ internal/models/currency.go
@@ -6,7 +6,7 @@ func (m *DBModel) GetCurrencies() (currencies []Currency, err error) {
 
 	query := `SELECT  
 			uuid, name, metric, target_value, time_frame_unit,
-			time_frame_value, time_frame_since, comparison, filters
+			time_frame_value, time_frame_since, comparison, filters, IFNULL(sub_metrics, '')
 		FROM currency`
 	rows, err := m.DB.QueryContext(ctx, query)
 	if err != nil {
@@ -17,7 +17,7 @@ func (m *DBModel) GetCurrencies() (currencies []Currency, err error) {
 	for rows.Next() {
 		var c Currency
 		if err = rows.Scan(&c.UUID, &c.Name, &c.Metric, &c.TargetValue, &c.TimeFrame.Unit,
-			&c.TimeFrame.Value, &c.TimeFrame.Since, &c.Comparison, &c.Filters); err != nil {
+			&c.TimeFrame.Value, &c.TimeFrame.Since, &c.Comparison, &c.Filters, &c.SubMetrics); err != nil {
 			return currencies, err
 		}
 		currencies = append(currencies, c)
@@ -32,11 +32,11 @@ func (m *DBModel) GetCurrency(uuid string) (c Currency, err error) {
 
 	query := `SELECT  
 			uuid, name, metric, target_value, time_frame_unit,
-			time_frame_value, time_frame_since, comparison, filters
+			time_frame_value, time_frame_since, comparison, filters, IFNULL(sub_metrics, '')
 		FROM currency WHERE uuid = ?`
 	row := m.DB.QueryRowContext(ctx, query, uuid)
 	if err = row.Scan(&c.UUID, &c.Name, &c.Metric, &c.TargetValue, &c.TimeFrame.Unit,
-		&c.TimeFrame.Value, &c.TimeFrame.Since, &c.Comparison, &c.Filters); err != nil {
+		&c.TimeFrame.Value, &c.TimeFrame.Since, &c.Comparison, &c.Filters, &c.SubMetrics); err != nil {
 		return c, err
 	}
 
@@ -49,10 +49,11 @@ func (m *DBModel) UpdateCurrency(c Currency) (err error) {
 
 	query := `UPDATE currency SET 
 			name = ?, metric = ?, target_value = ?, time_frame_unit = ?,
-			time_frame_value = ?, time_frame_since = ?, comparison = ?, filters = ?
+			time_frame_value = ?, time_frame_since = ?, comparison = ?, filters = ?,
+			sub_metrics = ?
 		WHERE uuid = ?`
 	_, err = m.DB.ExecContext(ctx, query, c.Name, c.Metric, c.TargetValue, c.TimeFrame.Unit,
-		c.TimeFrame.Value, c.TimeFrame.Since, c.Comparison, c.Filters, c.UUID)
+		c.TimeFrame.Value, c.TimeFrame.Since, c.Comparison, c.Filters, c.SubMetrics, c.UUID)
 	return err
 }
 
@@ -61,10 +62,10 @@ func (m *DBModel) InsertCurrency(c Currency) (err error) {
 	defer cancel()
 
 	query := `INSERT INTO currency (uuid, name, metric, target_value,
-			time_frame_unit, time_frame_value, time_frame_since, comparison, filters)
-		VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)`
+			time_frame_unit, time_frame_value, time_frame_since, comparison, filters, sub_metrics)
+		VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
 	_, err = m.DB.ExecContext(ctx, query, c.UUID, c.Name, c.Metric, c.TargetValue,
-		c.TimeFrame.Unit, c.TimeFrame.Value, c.TimeFrame.Since, c.Comparison, c.Filters)
+		c.TimeFrame.Unit, c.TimeFrame.Value, c.TimeFrame.Since, c.Comparison, c.Filters, c.SubMetrics)
 	return err
 }